Separate the event class and the actual event instance happening in the game
