+++ docs/Javascript/start-up.js
@@ -4,7 +4,7 @@
 import { display_ending_option_button, display_replay_button, display_attribution, play_area_direction_row } from "./imported_modules/tools.js";
 import { director } from "./imported_modules/main.js";
 import { update_play_area } from "./imported_modules/main.js";
-import * as event_file from "./imported_modules/event.js";
+import * as event_file from "./imported_modules/story.js";
 import * as config from "./imported_modules/config.js";
 
 const replay_button = document.querySelector(".replay-button");
